// src/app/interfaces/hero/i-hero-buildings.ts
export interface IHeroBuildings {
    agora: number;
    lumberMill: number;
    farm: number;
    tradeRoute: number;
    armory: number;
    barracks: number;
    fortress: number;
    argyroeides: number;
    quarry: number;
    oikia: number;
    oracle: number;
    temple: number;
    port: number;
    academy: number;
    gymnasium: number;
    fountain: number;
    altar: number;
    planningHall: number;
    chamber: number;
    royalPalace: number;
  }